refactor(settings): extract AgentPanel and ProvidersPanel out of SettingsPage
- AgentPanel (zero props, store-driven): framework card descriptors,
  install-state grouping, section-level Re-detect, and the uninstall /
  framework-switch dialogs with all their state
- ProvidersPanel: Active model / Reasoning effort / Providers sections,
  connection-test and Codex sign-in logic; takes only the form
  navigation callbacks and the shared busy-provider state
- SettingsPage drops from 1179 to 758 lines and keeps only the nav
  shell, history, provider form, and the open-time auto-detect effects
- Behavior note: a pending isolated Codex sign-in is now cancelled when
  the panel unmounts (dialog close or panel switch), not only on close
